float占用运算太多，修改一下算法，用int代替

+++ paperchecker/src/main/java/com.jack/utils/HammingUtils.java
@@ -31,7 +31,7 @@ public class HammingUtils {
      */
     public static double getSimilarity(String simHash1, String simHash2) {
         // 通过 simHash1 和 simHash2 获得它们的海明距离
-        float distance = getHammingDistance(simHash1, simHash2);
+        int distance = getHammingDistance(simHash1, simHash2);
         // 通过海明距离计算出相似度，并返回
         return 0.01 * (100 - distance * 100 / 128);
     }
